Move dice rolling into its own file
